Commit order before emails; use big integer total
CheckoutComponent: commit the DB transaction before sending emails and wrap Mail::send calls in a try/catch to log failures so email errors won't roll back the created order. Also removed the previous commented mail calls and keep cart clearing/event dispatching intact. Migration: change orders.total from unsignedTinyInteger to unsignedBigInteger to support larger order totals (run migrations after pulling this change).

+++ database/migrations/2026_04_30_084723_create_orders_table.php
@@ -17,7 +17,7 @@ return new class extends Migration {
             $table->string('email');
             $table->text('note')->nullable();
             $table->boolean('status')->default(false);
-            $table->unsignedTinyInteger('total');
+            $table->unsignedBigInteger('total');
             $table->timestamps();
         });
     }
